backend/requirements.txt
# Core Framework
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# Database
sqlalchemy==2.0.29
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# PDF / OCR / Tables
camelot-py[cv]==0.11.0
pdfplumber==0.10.3
pdf2image==1.17.0
pytesseract==0.3.10
opencv-python-headless==4.9.0.80
Pillow==10.2.0

# ML / Scoring
xgboost==2.0.3
scikit-learn==1.4.1.post1
shap==0.45.0
pandas==2.2.1
numpy==1.26.4
joblib==1.3.2
matplotlib==3.8.3

# Report Generation
python-docx==1.1.0
docx2pdf==0.1.8
reportlab==4.1.0
jinja2==3.1.3

# HTTP / APIs
httpx==0.27.0
aiohttp==3.9.3
python-dotenv==1.0.1

# Utility
pydantic==2.6.4
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
